Use Xeno XML parser again.
Dodge billion laughs attack.
